+++ src/eventNames.js
@@ -12,7 +12,7 @@ module.exports = {
   INIT: "init",
   START: "start",
   STOP: "stop",
-  LOAD_MOCKS: "load:mocks",
+  LOAD_LEGACY_MOCKS: "load:mocks:legacy",
   CHANGE_FIXTURES: "change:fixtures",
   CHANGE_MOCKS: "change:mocks",
   CHANGE_SETTINGS: "change:settings",