+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Remove "booleanString" option type. Now only "number", "boolean" or "string" can be used.
 - feat: Remove deprecated "onLoadMocks" method, "onChangeMocks" must be used instead.
 - feat: Remove "onLoadFiles" method. There is no alternative, as it is an internal event of the files-loader plugin and it should't be used by other external pieces.
+- feat: Rename "loadMocks" plugins method to "loadLegacyMocks"
 
 ## [1.6.0] - 2020-12-25